ENH mapreduce: correctly iterate over workers list to remove elements.
Removing elements in a list while iterating leads to undefined behaviour. While it was not causing bugs (processes finished completely before exit) it lead to believe that some processes are finished while they are not.

See https://docs.python.org/2/tutorial/datastructures.html#looping-techniques and the many questions on the internet for more explanation.